ability to launch file://

+++ R/RCCRF.R
@@ -6,30 +6,30 @@
 #' @return a list of functions for plumber to reference
 #' @export
 RCCRF <- function(config) {
-  config$addConfig(
-    RCCRFApi = list(label = "REDCap CRF API URL", type="url", required=TRUE),
-    RCCRFApiKey = list(label = "REDCap CRF Project API Key", type="password")
-  )
-  return(list(updateCRF = RCupdateCRF, crfDelete = RCCRFDelete))
+  return(list(updateCRF = RCupdateCRF, crfDelete = RCCRFDelete, getPDF = RCgetPDF))
 }
 #' Upload CRFs
 #'
-#' @param id A vector (length > 0) row ids in the sqlite database to be uploaded
+#' @param sn A vector (length > 0) cartridge_sn in the sqlite database to be uploaded
 #' @param config Scrapert config
 #'
 #' @return tibble 2 cols (id, uploaded) with the sqlite id, and an uploaded timestamp (UCT time POSIXct)
 #' @export
-RCupdateCRF <- function(id, config) {
+RCupdateCRF <- function(sn, config) {
   key <- RCCRFgetKey(config)
   # Build data frames
   con <- dbConnect(RSQLite::SQLite(), "xpertdb")
-  df <- tbl(con, "xpert_results") %>% filter(id %in% local(id)) %>% 
-    select(id, sample_id, cartridge_sn, test_result, end_time, processed_by) %>% collect() %>%
-    # Get the record numbers
-    RCCRFgetRecordId(config, key)
+  df <- tbl(con, "xpert_results") %>% filter(cartridge_sn %in% local(sn) & !is.na(pid)) %>% 
+    select(id, sample_id, cartridge_sn, test_result, end_time, processed_by, pid) %>% collect()
+  if (nrow(df) == 0) {
+    dbDisconnect(con);
+    stop (sprintf("pid is missing for %s", sn))
+  };
+  df <- df %>% # Get the record numbers
+    RCCRFgetRecordId(config, key) %>% filter(!is.na(record_id))
   # upload xpert results
   data <- list(
-    token = keyring::key_get("Scrapert-RCCRFApiKey"),
+    token = key,
     content = "record",
     format = "csv",
     type = "flat",
@@ -41,51 +41,60 @@ RCupdateCRF <- function(id, config) {
                                            cartridge_sn, xpert_processed_by = processed_by, -sample_id, -id) %>%
                                mutate(xpert_result_complete = 2))
   )
-  response <- httr::POST(config$config$RCCRFApi$value, body = data, encode = "form")
+  response <- httr::POST(config$config$api, body = data, encode = "form")
   if (response$status_code != 200) stop(paste("REDCap Error updating CRF data", response$status_code, httr::content(response)))
-  count <-as.integer(httr::content(response)$count)
-  if (count != length(id)) stop(paste("REDCap only updated", count, "records but updates to", length(id), "records was requested"))
   # Now for the files
   apply(df, MARGIN = 1, FUN = function(x) {
     x <- as.list(x)
-    temp <- file(paste0(tempdir(), "/", x$sample_id, ".pdf"), "wb")
-    base64enc::base64decode(tbl(con, "xpert_results") %>% filter(id == local(x$id)) %>% select(pdf) %>% collect() %>%
-                            pull(pdf), output = temp)
-    flush(temp)
-    temppath <- summary(temp)$description
-    data = list(
-      token = key,
-      action = "import",
-      content = "file",
-      record = x$record_id,
-      field = "xpert_data",
-      event= "",
-      returnFormat = "json",
-      file = httr::upload_file(temppath)
-    )
-    response <- httr::POST(config$config$RCCRFApi$value, body = data, encode = "multipart")
-    if (response$status_code != 200) stop(paste("REDCap File Upload Error", response$status_code, httr::content(response)))
-    close(temp)
-    file.remove(temppath)
+    pdf <- tbl(con, "xpert_results") %>% filter(id == local(x$id)) %>% select(pdf) %>% collect() %>%
+      pull(pdf)
+    if (!is.na(pdf)) {
+      temp <- file(paste0(tempdir(), "/", x$sample_id, ".pdf"), "wb")
+      base64enc::base64decode(pdf, output = temp)
+      flush(temp)
+      temppath <- summary(temp)$description
+      data = list(
+        token = key,
+        action = "import",
+        content = "file",
+        record = x$record_id,
+        field = "xpert_data",
+        event= "",
+        returnFormat = "json",
+        file = httr::upload_file(temppath)
+      ) 
+      response <- httr::POST(config$config$api, body = data, encode = "multipart")
+      if (response$status_code != 200) stop(paste("REDCap File Upload Error", response$status_code, httr::content(response)))
+      close(temp)
+      file.remove(temppath)
+      # remove from the database
+      if(con %>% dbExecute(paste("UPDATE xpert_results SET pdf = NULL WHERE id =", x$id)) != 1) 
+        stop("Error removing pdf from local database")
+    } else logger::log_warn("{ x$cartridge_sn } pdf missing from sqlite3, perhpas it's uploaded already")
+    pdf <- NULL
   })
   dbDisconnect(con)
-  df$id
+  df$cartridge_sn
 }
 
 #' REDCap Delete CRF
 #'
-#' @param id the row id in xpertdb
+#' @param sn the cartridge_sn in xpertdb
 #' @param config \link{Config} object
 #'
 #' @return \code{NULL} if successful
 #' @export
-RCCRFDelete <- function(id, config) {
+RCCRFDelete <- function(sn, config) {
   key <- RCCRFgetKey(config)
   con <- dbConnect(RSQLite::SQLite(), "xpertdb")
-  df <- tbl(con, "xpert_results") %>% filter(id %in% local(id)) %>% 
-    select(id, sample_id) %>% collect() %>% 
+  df <- tbl(con, "xpert_results") %>% filter(cartridge_sn %in% local(sn)) %>% 
+    select(pid) %>% collect()
+  if (is.na(df$pid)) {
+    logger::log_info("No PID set for sn {sn}, unable to delete from REDCap")
+    return()
+  }
     # get the record numbers
-    RCCRFgetRecordId(config, key)
+  df <- df %>% RCCRFgetRecordId(config, key)
   # delete the files first
   dbDisconnect(con)
   data <- list(
@@ -97,11 +106,11 @@ RCCRFDelete <- function(id, config) {
     event= "",
     returnFormat = "json"
   )
-  response <- httr::POST(config$config$RCCRFApi$value, body = data, encode = "form")
-  if (response$status_code != 200) stop(paste("REDCap File delete Error", response$status_code, httr::content(response)))
+  response <- httr::POST(config$config$api, body = data, encode = "form")
+  if (response$status_code != 200) logger::log_warn("REDCap File delete Error { response$status_code }- {httr::content(response , as='text')}")
   # We don't delete just remove values
   data <- list(
-    token = keyring::key_get("Scrapert-RCCRFApiKey"),
+    token = key,
     content = "record",
     format = "csv",
     type = "flat",
@@ -109,14 +118,14 @@ RCCRFDelete <- function(id, config) {
     forceAutoNumber = "false",
     returnContent = "count",
     returnFormat = "json",
-    data = readr::format_csv(df %>% select(record_id, -sample_id, -id) %>%
+    data = readr::format_csv(df %>% select(record_id, -pid) %>%
                                mutate(xpert_result_complete = "", xpert_result = "", xpert_timestamp = "",
                                       cartridge_sn = "", xpert_processed_by = ""))
   )
-  response <- httr::POST(config$config$RCCRFApi$value, body = data, encode = "form")
+  response <- httr::POST(config$config$api, body = data, encode = "form")
   if (response$status_code != 200) stop(paste("REDCap Error deleting CRF data", response$status_code, httr::content(response)))
   count <-as.integer(httr::content(response)$count)
-  if (count != length(id)) stop(paste("REDCap only deleted", count, "records but deletion ", length(id), "records was requested"))
+  if (count != length(sn)) stop(paste("REDCap only deleted", count, "records but deletion ", length(sn), "records was requested"))
   NULL
 }
 
@@ -128,13 +137,31 @@ RCCRFDelete <- function(id, config) {
 #' @return API Key
 RCCRFgetKey <- function (config) {
   key <- tryCatch({
-    keyring::key_get("Scrapert-RCCRFApiKey")
+    keyring::key_get("Scrapert-apikey")
   }, error = function(e) {
     return(NULL)
   })
-  if (!is.character(key)) stop("REDCap CRF API Key is not set")
-  if (!is.character(config$config$RCCRFApi$value)) stop("REDCap CRF Api URL is not set")
+  if (!is.character(key)) stop("REDCap API Key is not set")
+  if (!is.character(config$config$api)) stop("REDCap Api URL is not set")
   key
+}
+
+RCgetPDF <- function(pid, config) {
+  key = RCCRFgetKey(config)
+  df <- RCCRFgetRecordId(tibble(pid = pid), config, key)
+  data <- list(
+    token = key,
+    content = "file",
+    action = "export",
+    record = df$record_id,
+    field = "xpert_data",
+    event = "",
+    returnFormat = "json"
+  )
+  response <- httr::POST(config$config$api, body = data, encode = "form")
+  if (response$status_code != 200) stop(paste("REDCap Error getting pdf", response$status_code, httr::content(response)))
+  pdf <- httr::content(response, as="raw")
+  pdf
 }
 
 #' REDCap Map row id to record id
@@ -151,11 +178,11 @@ RCCRFgetRecordId <- function(df, config, key) {
     format = "csv",
     type = "flat",
     returnFormat = "json",
-    fields = "record_id, sample_id",
-    filterLogic = paste(paste0("[sample_id]='", df$sample_id, "'"), collapse = " OR ")
+    fields = "record_id, pid",
+    filterLogic = paste(paste0("[pid]='", df$pid, "'"), collapse = " OR ")
   )
-  response <- httr::POST(config$config$RCCRFApi$value, body = data, encode = "form")
+  response <- httr::POST(config$config$api, body = data, encode = "form")
   if (response$status_code != 200) stop(paste("REDCap Error getting record_id", response$status_code, httr::content(response)))
-  rid <- httr::content(response, show_col_types = FALSE)
-  df %>% left_join(rid, by="sample_id")
+  rid <- httr::content(response, col_types="cc")
+  df %>% left_join(rid, by="pid")
 }
